flask-server/requirements.txt
flask
flask_restful
python-dotenv
psycopg2